app/requirements.txt
langchain==0.0.304
protobuf==3.20.3
PyAudio==0.2.13
pydantic==2.4.2
python-dotenv==1.0.0
recorder==0.0.2
squareup==30.0.0.20230816
google-cloud-texttospeech==2.14.1
chromadb